requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Django==4.0.4
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-WTF==1.0.1
fonttools==4.33.3
image==1.5.33
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0

requests~=2.27.1
colorthief~=0.2.1